First iteration fixes
