+++ server.js
@@ -6,5 +6,5 @@ const PORT = process.env.PORT || 3000;
 app.use(express.static('./dist'));
 
 app.listen(PORT, function () {
-  console.log(`Example app listening on port ${PORT}!`);
-}); 
+  console.log('Example app is here http://localhost:3000/');
+});
