+++ src/components/Chat/Chat.hbs
@@ -18,4 +18,5 @@
         {{/if}}
     </div>
     {{#if isChatSelected}}{{{ChatMessageInput}}}{{/if}}
+    {{{Modal}}}
 </div>
