+++ src/components/ChatMessage/ChatMessage.hbs
@@ -42,7 +42,7 @@
             {{#if isMessageSent}}
               {{#if isMessageDelivered}}
                 <img
-                    src="../../../static/images/messageIsReceivedIcon.png"
+                    src={{messageReceivedImgPath}}
                     alt="Знак получения сообщения"
                     class="chatMessage__messageIsReceivedIcon"
                 >
