+++ src/components/Input/Input.hbs
@@ -1,5 +1,5 @@
 {{#if isTextArea}}
 <textarea {{#if value}}value={{value}}{{/if}} class="{{inputStyle}}" name={{name}} placeholder={{placeholder}}>
 {{else}}
-<input type="{{type}}" {{#if value}}value={{value}}{{/if}} class="{{inputStyle}}" name={{name}} id={{id}}>
+<input type="{{type}}" {{#if value}}value={{value}}{{/if}} class="{{inputStyle}}" name={{name}} {{#if id}}id={{id}}{{/if}}>
 {{/if}}
