+++ tsconfig.json
@@ -25,7 +25,5 @@
     "noUnusedParameters": true,
     "strictNullChecks": true
   },
-  "include": [
-    "src"
-  ]
+  "include": ["src"],
 }
